Cambio de Amount de Int a Double

+++ Flutter/example_superapp/android/src/main/kotlin/com/genexus/example_superapp/ISuperApp.kt
@@ -4,8 +4,8 @@ import com.genexus.android.core.base.model.Entity
 import io.flutter.plugin.common.MethodChannel
 
 interface ISuperApp {
-	fun payWithoutUI(amount: Int, resultHandler: MethodChannel.Result? = null)
-	fun payWithUI(amount: Int, resultHandler: MethodChannel.Result? = null)
+	fun payWithoutUI(amount: Double, resultHandler: MethodChannel.Result? = null)
+	fun payWithUI(amount: Double, resultHandler: MethodChannel.Result? = null)
 	fun getClientInformation(clientId: String, resultHandler: MethodChannel.Result? = null)
 	fun getPaymentInformation(clientInformation: Entity, resultHandler: MethodChannel.Result? = null)
 	fun getPaymentInformation(clientInformation: HashMap<String, String>, resultHandler: MethodChannel.Result? = null)
